Use custom value for Let's Encrypt creds
TL;DR
-----

Set up a custom property for Let's Encrypt credentials rather
that re-using the kpack credentials

Details
-------

The previous approach to connecting to GCP for the DNS solver
for Let's Encrypt re-used the parameters passed to kpack for
accessing `gcr.io`. This didn't feel right, so it's now split
out into its own value at `certmanager.solver.clouddns.keyfile`.

A default empty value is set by the overlays, and then the
`setup` script overwrites it using an environment variable
based on the value passed in.